moves firebase.js and UserContext.js to resources directory
