+++ src/components/Header.js
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from 'react';
 import { Link } from 'react-router-dom';
-import firebase from '../config/firebase';
+import firebase from '../resources/firebase';
 
 const Header = () => {
     const [profilePictureUrl, setProfilePictureUrl] = useState(null);
