+++ src/resources/firebase.js
